third_party/WebKit/LayoutTests/fast/dom/gc-11.html

<!DOCTYPE html>
<script src="../js/resources/js-test-pre.js"></script>
<script>

description('This test checks that custom data on reachable DOM nodes is not lost during GC.');

var xmlDoc = document.implementation.createDocument('', 'doc', null);
var xmlNode = xmlDoc.documentElement;
xmlNode.ownerDocument.data = 1;
xmlDoc = null;

gc();

shouldBe('xmlNode.ownerDocument.data', '1');

</script>
<script src="../js/resources/js-test-post.js"></script>
